// src/app/transfers/models/api/partnerAccountsModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PartnerAccount } from './partnerAccount';

export interface PartnerAccountsModel { 
    partnerId?: string;
    accounts?: Array<PartnerAccount>;
}